gastronomy: move the permit/concession machinery down from enigmatic
`Permit`/`ProcessingPermit`/`Concession` and the `crypto.permit…Crypto`
aggregates now live in gastronomy, shared by both modules: gastronomy gates
its own weak hashes (MD5 disallowed, SHA-1 deprecated) on `digest` and
`checksum`, while enigmatic keeps only the cipher-side `Weakness` /
`Authentication` match types and imports the rest.

`Concession` holds every marker, hash and cipher alike — the cipher ones
are only used downstream by enigmatic — so the named aggregate permits can
be defined once, in the lowest crypto module. The HMAC extension and
`BloomFilter` propagate the new weak-hash permit to their callers.
