+++ lib/enigmatic/src/core/enigmatic_expose.scala
@@ -38,6 +38,7 @@ import javax.crypto as jc
 import anticipation.*
 import contingency.*
 import distillate.*
+import gastronomy.{Permit, ProcessingPermit}
 import gossamer.*
 import prepositional.*
 import vacuous.*
