+++ lib/enigmatic/src/core/soundness_enigmatic_core.scala
@@ -32,13 +32,15 @@
                                                                                                   */
 package soundness
 
+// `Concession`, `Permit`, `ProcessingPermit` and the `crypto.permit…Crypto`
+// aggregates are re-exported by gastronomy (where they now live).
 export
   enigmatic
   . { Aes, Blowfish, BlockCipher, BlockCipherMode, BlockCipherPadding, Cbc, Cfb, Cipher,
-      CipherSession, Concession, Crypto, CryptoError, Ctr, decrypt, Decryptor, Des, Divulgence,
+      CipherSession, Crypto, CryptoError, Ctr, decrypt, Decryptor, Des, Divulgence,
       Dsa, Ecb, encrypt, Encryptor, Encryption, expose,
-      Hmac, hmac, InitializationVector, Iso10126, NoPadding, Ofb, Pem, PemError, PemLabel, Permit,
-      Permits, Pkcs7, PrivateKey, ProcessingPermit, PublicKey, Rc2, Rsa, Signature, Signing,
+      Hmac, hmac, InitializationVector, Iso10126, NoPadding, Ofb, Pem, PemError, PemLabel,
+      Permits, Pkcs7, PrivateKey, PublicKey, Rc2, Rsa, Signature, Signing,
       Symmetric, SymmetricKey, TripleDes }
 
 package blockCipherMode:
@@ -54,7 +56,3 @@ package initializationVector:
 
 package cryptoProviders:
   export enigmatic.cryptoProviders.javaStdlibCrypto
-
-package crypto:
-  export enigmatic.crypto.{permitUnauthenticatedCrypto, permitDeprecatedCrypto, permitLegacyCrypto,
-      permitDisallowedCrypto}
