+++ lib/ulysses/src/core/ulysses.BloomFilter.scala
@@ -45,7 +45,7 @@ import vacuous.*
 object BloomFilter:
   def apply[element: Digestible](approximateSize: Int, targetErrorRate: 0.0 ~ 1.0)
     [ algorithm <: Algorithm ]
-    ( using Hash in algorithm )
+    ( using hash0: Hash in algorithm, erased weakness: Permit[HashWeakness[algorithm]] )
   :   BloomFilter[element, algorithm] =
 
     val bitSize: Int = (-1.44*approximateSize*ln(targetErrorRate.double).double).toInt
@@ -55,7 +55,7 @@ object BloomFilter:
 
 case class BloomFilter[element: Digestible, algorithm <: Algorithm]
   ( bitSize: Int, hashCount: Int, bits: sci.BitSet )
-  ( using Hash in algorithm ):
+  ( using hash0: Hash in algorithm, erased weakness: Permit[HashWeakness[algorithm]] ):
 
   private val requiredEntropyBits = ln(bitSize ** hashCount).double.toInt + 1
